Fix ContextualSaveBar circular deps
