+++ .eslintrc
@@ -28,7 +28,6 @@
         "allowBlockStart": false
       }
     ],
-    "import/no-cycle": "off",
     "import/no-named-as-default": "off",
     "react/button-has-type": "off",
     "react/no-array-index-key": "off",
