+++ src/components/ContextualSaveBar/index.ts
@@ -1,4 +1,3 @@
 import ContextualSaveBar from './ContextualSaveBar';
 
 export default ContextualSaveBar;
-export {Props as ContextualSaveBarProps} from './ContextualSaveBar';
